requirements.txt
dbt-core==1.7.14
dbt-bigquery==1.7.6
google-cloud-bigquery==3.11.0
apache-airflow==2.9.2
boto3==1.28.32
python-dotenv==1.0.1
pandas==2.2.3
scikit-learn==1.3.2
statsmodels==0.14.2
matplotlib==3.7.2
joblib==1.3.2
requests==2.32.3
